feat: add UserDefaults to save users todo list

+++ TodoList/Views/ListView.swift
@@ -11,23 +11,37 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
 
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItem(item: item)
-                        }
+        Group {
+            if listViewModel.items.isEmpty {
+                VStack(alignment: .center, spacing: 8) {
+                    Text("Your list is empty")
+                        .fontWeight(.semibold)
+
+                    Text("Tap Add to insert your first todo item")
+                        .foregroundColor(.secondary)
+                }
+                .padding()
+                .multilineTextAlignment(.center)
+
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+                .listStyle(.plain)
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
         }
         .navigationTitle("Todo List")
         .navigationBarItems(leading: EditButton(),
                             trailing: NavigationLink("Add", destination: AddView()))
-
-        .listStyle(.plain)
     }
 }
